+++ crates/chain/src/indexer/keychain_txout.rs
@@ -783,74 +783,6 @@ impl<K: Clone + Ord + Debug> KeychainTxOutIndex<K> {
     }
 }
 
-#[cfg(feature = "sqlite")]
-impl ChangeSet {
-    /// Schema name for the changeset.
-    pub const SCHEMA_NAME: &'static str = "bdk_keychaintxout";
-    /// Name for table that stores last revealed indices per descriptor id.
-    pub const LAST_REVEALED_TABLE_NAME: &'static str = "bdk_descriptor_last_revealed";
-
-    /// Initialize sqlite tables for persisting [`KeychainTxOutIndex`].
-    fn init_sqlite_tables(db_tx: &rusqlite::Transaction) -> rusqlite::Result<()> {
-        let schema_v0: &[&str] = &[
-            // last revealed
-            &format!(
-                "CREATE TABLE {} ( \
-                descriptor_id TEXT PRIMARY KEY NOT NULL, \
-                last_revealed INTEGER NOT NULL \
-                ) STRICT",
-                Self::LAST_REVEALED_TABLE_NAME,
-            ),
-        ];
-        crate::sqlite::migrate_schema(db_tx, Self::SCHEMA_NAME, &[schema_v0])
-    }
-
-    /// Construct [`KeychainTxOutIndex`] from sqlite database and given parameters.
-    pub fn from_sqlite(db_tx: &rusqlite::Transaction) -> rusqlite::Result<Self> {
-        Self::init_sqlite_tables(db_tx)?;
-        use crate::sqlite::Sql;
-
-        let mut changeset = Self::default();
-
-        let mut statement = db_tx.prepare(&format!(
-            "SELECT descriptor_id, last_revealed FROM {}",
-            Self::LAST_REVEALED_TABLE_NAME,
-        ))?;
-        let row_iter = statement.query_map([], |row| {
-            Ok((
-                row.get::<_, Sql<DescriptorId>>("descriptor_id")?,
-                row.get::<_, u32>("last_revealed")?,
-            ))
-        })?;
-        for row in row_iter {
-            let (Sql(descriptor_id), last_revealed) = row?;
-            changeset.last_revealed.insert(descriptor_id, last_revealed);
-        }
-
-        Ok(changeset)
-    }
-
-    /// Persist `changeset` to the sqlite database.
-    pub fn persist_to_sqlite(&self, db_tx: &rusqlite::Transaction) -> rusqlite::Result<()> {
-        Self::init_sqlite_tables(db_tx)?;
-        use crate::rusqlite::named_params;
-        use crate::sqlite::Sql;
-
-        let mut statement = db_tx.prepare_cached(&format!(
-            "REPLACE INTO {}(descriptor_id, last_revealed) VALUES(:descriptor_id, :last_revealed)",
-            Self::LAST_REVEALED_TABLE_NAME,
-        ))?;
-        for (&descriptor_id, &last_revealed) in &self.last_revealed {
-            statement.execute(named_params! {
-                ":descriptor_id": Sql(descriptor_id),
-                ":last_revealed": last_revealed,
-            })?;
-        }
-
-        Ok(())
-    }
-}
-
 #[derive(Clone, Debug, PartialEq)]
 /// Error returned from [`KeychainTxOutIndex::insert_descriptor`]
 pub enum InsertDescriptorError<K> {
