+++ crates/chain/src/local_chain.rs
@@ -681,83 +681,6 @@ impl FromIterator<(u32, BlockHash)> for ChangeSet {
     }
 }
 
-#[cfg(feature = "sqlite")]
-impl ChangeSet {
-    /// Schema name for the changeset.
-    pub const SCHEMA_NAME: &'static str = "bdk_localchain";
-    /// Name of sqlite table that stores blocks of [`LocalChain`].
-    pub const BLOCKS_TABLE_NAME: &'static str = "bdk_blocks";
-
-    /// Initialize sqlite tables for persisting [`LocalChain`].
-    fn init_sqlite_tables(db_tx: &rusqlite::Transaction) -> rusqlite::Result<()> {
-        let schema_v0: &[&str] = &[
-            // blocks
-            &format!(
-                "CREATE TABLE {} ( \
-                block_height INTEGER PRIMARY KEY NOT NULL, \
-                block_hash TEXT NOT NULL \
-                ) STRICT",
-                Self::BLOCKS_TABLE_NAME,
-            ),
-        ];
-        crate::sqlite::migrate_schema(db_tx, Self::SCHEMA_NAME, &[schema_v0])
-    }
-
-    /// Construct a [`LocalChain`] from sqlite database.
-    pub fn from_sqlite(db_tx: &rusqlite::Transaction) -> rusqlite::Result<Self> {
-        Self::init_sqlite_tables(db_tx)?;
-        use crate::sqlite::Sql;
-
-        let mut changeset = Self::default();
-
-        let mut statement = db_tx.prepare(&format!(
-            "SELECT block_height, block_hash FROM {}",
-            Self::BLOCKS_TABLE_NAME,
-        ))?;
-        let row_iter = statement.query_map([], |row| {
-            Ok((
-                row.get::<_, u32>("block_height")?,
-                row.get::<_, Sql<BlockHash>>("block_hash")?,
-            ))
-        })?;
-        for row in row_iter {
-            let (height, Sql(hash)) = row?;
-            changeset.blocks.insert(height, Some(hash));
-        }
-
-        Ok(changeset)
-    }
-
-    /// Persist `changeset` to the sqlite database.
-    pub fn persist_to_sqlite(&self, db_tx: &rusqlite::Transaction) -> rusqlite::Result<()> {
-        Self::init_sqlite_tables(db_tx)?;
-        use crate::sqlite::Sql;
-        use rusqlite::named_params;
-
-        let mut replace_statement = db_tx.prepare_cached(&format!(
-            "REPLACE INTO {}(block_height, block_hash) VALUES(:block_height, :block_hash)",
-            Self::BLOCKS_TABLE_NAME,
-        ))?;
-        let mut delete_statement = db_tx.prepare_cached(&format!(
-            "DELETE FROM {} WHERE block_height=:block_height",
-            Self::BLOCKS_TABLE_NAME,
-        ))?;
-        for (&height, &hash) in &self.blocks {
-            match hash {
-                Some(hash) => replace_statement.execute(named_params! {
-                    ":block_height": height,
-                    ":block_hash": Sql(hash),
-                })?,
-                None => delete_statement.execute(named_params! {
-                    ":block_height": height,
-                })?,
-            };
-        }
-
-        Ok(())
-    }
-}
-
 /// An error which occurs when a [`LocalChain`] is constructed without a genesis checkpoint.
 #[derive(Clone, Debug, PartialEq)]
 pub struct MissingGenesisError;
