+++ crates/chain/src/tx_graph.rs
@@ -1293,188 +1293,6 @@ impl<A> ChangeSet<A> {
     }
 }
 
-#[cfg(feature = "sqlite")]
-impl<A> ChangeSet<A>
-where
-    A: Anchor + Clone + Ord + serde::Serialize + serde::de::DeserializeOwned,
-{
-    /// Schema name for the [`ChangeSet`].
-    pub const SCHEMA_NAME: &'static str = "bdk_txgraph";
-    /// Name of table that stores full transactions and `last_seen` timestamps.
-    pub const TXS_TABLE_NAME: &'static str = "bdk_txs";
-    /// Name of table that stores floating txouts.
-    pub const TXOUTS_TABLE_NAME: &'static str = "bdk_txouts";
-    /// Name of table that stores [`Anchor`]s.
-    pub const ANCHORS_TABLE_NAME: &'static str = "bdk_anchors";
-
-    /// Initialize sqlite tables.
-    fn init_sqlite_tables(db_tx: &rusqlite::Transaction) -> rusqlite::Result<()> {
-        let schema_v0: &[&str] = &[
-            // full transactions
-            &format!(
-                "CREATE TABLE {} ( \
-                txid TEXT PRIMARY KEY NOT NULL, \
-                raw_tx BLOB, \
-                last_seen INTEGER \
-                ) STRICT",
-                Self::TXS_TABLE_NAME,
-            ),
-            // floating txouts
-            &format!(
-                "CREATE TABLE {} ( \
-                txid TEXT NOT NULL, \
-                vout INTEGER NOT NULL, \
-                value INTEGER NOT NULL, \
-                script BLOB NOT NULL, \
-                PRIMARY KEY (txid, vout) \
-                ) STRICT",
-                Self::TXOUTS_TABLE_NAME,
-            ),
-            // anchors
-            &format!(
-                "CREATE TABLE {} ( \
-                txid TEXT NOT NULL REFERENCES {} (txid), \
-                block_height INTEGER NOT NULL, \
-                block_hash TEXT NOT NULL, \
-                anchor BLOB NOT NULL, \
-                PRIMARY KEY (txid, block_height, block_hash) \
-                ) STRICT",
-                Self::ANCHORS_TABLE_NAME,
-                Self::TXS_TABLE_NAME,
-            ),
-        ];
-        crate::sqlite::migrate_schema(db_tx, Self::SCHEMA_NAME, &[schema_v0])
-    }
-
-    /// Construct a [`TxGraph`] from an sqlite database.
-    pub fn from_sqlite(db_tx: &rusqlite::Transaction) -> rusqlite::Result<Self> {
-        Self::init_sqlite_tables(db_tx)?;
-        use crate::sqlite::Sql;
-
-        let mut changeset = Self::default();
-
-        let mut statement = db_tx.prepare(&format!(
-            "SELECT txid, raw_tx, last_seen FROM {}",
-            Self::TXS_TABLE_NAME,
-        ))?;
-        let row_iter = statement.query_map([], |row| {
-            Ok((
-                row.get::<_, Sql<Txid>>("txid")?,
-                row.get::<_, Option<Sql<Transaction>>>("raw_tx")?,
-                row.get::<_, Option<u64>>("last_seen")?,
-            ))
-        })?;
-        for row in row_iter {
-            let (Sql(txid), tx, last_seen) = row?;
-            if let Some(Sql(tx)) = tx {
-                changeset.txs.insert(Arc::new(tx));
-            }
-            if let Some(last_seen) = last_seen {
-                changeset.last_seen.insert(txid, last_seen);
-            }
-        }
-
-        let mut statement = db_tx.prepare(&format!(
-            "SELECT txid, vout, value, script FROM {}",
-            Self::TXOUTS_TABLE_NAME,
-        ))?;
-        let row_iter = statement.query_map([], |row| {
-            Ok((
-                row.get::<_, Sql<Txid>>("txid")?,
-                row.get::<_, u32>("vout")?,
-                row.get::<_, Sql<Amount>>("value")?,
-                row.get::<_, Sql<bitcoin::ScriptBuf>>("script")?,
-            ))
-        })?;
-        for row in row_iter {
-            let (Sql(txid), vout, Sql(value), Sql(script_pubkey)) = row?;
-            changeset.txouts.insert(
-                OutPoint { txid, vout },
-                TxOut {
-                    value,
-                    script_pubkey,
-                },
-            );
-        }
-
-        let mut statement = db_tx.prepare(&format!(
-            "SELECT json(anchor), txid FROM {}",
-            Self::ANCHORS_TABLE_NAME,
-        ))?;
-        let row_iter = statement.query_map([], |row| {
-            Ok((
-                row.get::<_, Sql<A>>("json(anchor)")?,
-                row.get::<_, Sql<Txid>>("txid")?,
-            ))
-        })?;
-        for row in row_iter {
-            let (Sql(anchor), Sql(txid)) = row?;
-            changeset.anchors.insert((anchor, txid));
-        }
-
-        Ok(changeset)
-    }
-
-    /// Persist `changeset` to the sqlite database.
-    pub fn persist_to_sqlite(&self, db_tx: &rusqlite::Transaction) -> rusqlite::Result<()> {
-        Self::init_sqlite_tables(db_tx)?;
-        use crate::rusqlite::named_params;
-        use crate::sqlite::Sql;
-
-        let mut statement = db_tx.prepare_cached(&format!(
-            "INSERT INTO {}(txid, raw_tx) VALUES(:txid, :raw_tx) ON CONFLICT(txid) DO UPDATE SET raw_tx=:raw_tx",
-            Self::TXS_TABLE_NAME,
-        ))?;
-        for tx in &self.txs {
-            statement.execute(named_params! {
-                ":txid": Sql(tx.compute_txid()),
-                ":raw_tx": Sql(tx.as_ref().clone()),
-            })?;
-        }
-
-        let mut statement = db_tx
-            .prepare_cached(&format!(
-                "INSERT INTO {}(txid, last_seen) VALUES(:txid, :last_seen) ON CONFLICT(txid) DO UPDATE SET last_seen=:last_seen",
-                Self::TXS_TABLE_NAME,
-            ))?;
-        for (&txid, &last_seen) in &self.last_seen {
-            statement.execute(named_params! {
-                ":txid": Sql(txid),
-                ":last_seen": Some(last_seen),
-            })?;
-        }
-
-        let mut statement = db_tx.prepare_cached(&format!(
-            "REPLACE INTO {}(txid, vout, value, script) VALUES(:txid, :vout, :value, :script)",
-            Self::TXOUTS_TABLE_NAME,
-        ))?;
-        for (op, txo) in &self.txouts {
-            statement.execute(named_params! {
-                ":txid": Sql(op.txid),
-                ":vout": op.vout,
-                ":value": Sql(txo.value),
-                ":script": Sql(txo.script_pubkey.clone()),
-            })?;
-        }
-
-        let mut statement = db_tx.prepare_cached(&format!(
-            "REPLACE INTO {}(txid, block_height, block_hash, anchor) VALUES(:txid, :block_height, :block_hash, jsonb(:anchor))",
-            Self::ANCHORS_TABLE_NAME,
-        ))?;
-        for (anchor, txid) in &self.anchors {
-            let anchor_block = anchor.anchor_block();
-            statement.execute(named_params! {
-                ":txid": Sql(*txid),
-                ":block_height": anchor_block.height,
-                ":block_hash": Sql(anchor_block.hash),
-                ":anchor": Sql(anchor.clone()),
-            })?;
-        }
-
-        Ok(())
-    }
-}
-
 impl<A: Ord> Merge for ChangeSet<A> {
     fn merge(&mut self, other: Self) {
         // We use `extend` instead of `BTreeMap::append` due to performance issues with `append`.
